Rename rotation.py to exit_scheduler.py
